chat_agent/requirements.txt
# Core — Microsoft Agent Framework
agent-framework-core>=0.1.0
botbuilder-core>=4.16.0
aiohttp>=3.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# Data
pandas>=2.2.0
openpyxl>=3.1.0
tabulate>=0.9.0
pyodbc>=5.0.0
